Relocate the accessibility package into the autofill package
