+++ app/src/main/java/com/x8bit/bitwarden/data/tiles/BitwardenAutofillTileService.kt
@@ -7,7 +7,7 @@ import android.os.Build
 import android.service.quicksettings.TileService
 import androidx.annotation.Keep
 import com.x8bit.bitwarden.AccessibilityActivity
-import com.x8bit.bitwarden.data.accessibility.manager.AccessibilityAutofillManager
+import com.x8bit.bitwarden.data.autofill.accessibility.manager.AccessibilityAutofillManager
 import com.x8bit.bitwarden.data.platform.annotation.OmitFromCoverage
 import com.x8bit.bitwarden.data.platform.util.isBuildVersionBelow
 import dagger.hilt.android.AndroidEntryPoint
